Implement Cyrillic to Latin conversion
In the process of implementing the reverse direction, several limitations were addressed in the rules. A lower error rate was observed on a small Wikipedia text corpus.

Notable changes include the removal of the vertical bar (|) for precedence. Instead, the bi-gram "kh" was introduced. Capital letters are now handled better. In Russian, ъ was mapped to ` since double quotes are commonly used in Slavic texts.

Closes #2.